exceptions

+++ src/main/java/com/renee/PhotoBlog/service/PhotosService.java
@@ -1,5 +1,6 @@
 package com.renee.PhotoBlog.service;
 
+import com.renee.PhotoBlog.exception.ResourceNotFoundException;
 import com.renee.PhotoBlog.model.Photo;
 import com.renee.PhotoBlog.repo.PhotosRepository;
 import lombok.extern.slf4j.Slf4j;
@@ -28,10 +29,14 @@ public class PhotosService {
     }
 
     public Optional<Photo> getPhotoById(Long id) {
-        return photosRepository.findById(id);
+        return Optional.ofNullable(photosRepository.findById(id)
+                .orElseThrow(() -> new ResourceNotFoundException("Photo with ID " + id + " not found")));
     }
 
     public void deletePhoto(Long id) {
+        if (!photosRepository.existsById(id)) {
+            throw new ResourceNotFoundException("Cannot delete, photo with ID " + id + " not found");
+        }
         photosRepository.deleteById(id);
     }
 }
